# spec/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Log Analyzer API
  version: 1.0.0
  summary: API for local-first log ingestion, analysis, incidents, and Ollama-powered insights.
  description: |
    This API is designed for a local backend that can be securely exposed over DuckDNS.
    It supports ingestion sources, parsed events, incident workflows, dashboard metrics,
    and AI analysis through Ollama with selectable model profiles.
servers:
  - url: https://your-subdomain.duckdns.org
    description: Public HTTPS endpoint behind reverse proxy
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sources:
    get:
      tags: [Sources]
      summary: List configured log sources
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Source'
                required: [items]
    post:
      tags: [Sources]
      summary: Create a source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreateRequest'
      responses:
        "201":
          description: Created source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
  /api/v1/sources/{id}:
    patch:
      tags: [Sources]
      summary: Update source
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcePatchRequest'
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
  /api/v1/sources/{id}/test:
    post:
      tags: [Sources]
      summary: Test source connectivity and parser compatibility
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  details:
                    type: string
                required: [ok]
...
  /api/v1/rules:
    get:
      tags: [Rules]
      summary: List detection rules
      responses:
        "200":
          description: Rule list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Rule'
                required: [items]
    post:
      tags: [Rules]
      summary: Create detection rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreateRequest'
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
  /api/v1/rules/{id}:
    patch:
      tags: [Rules]
      summary: Update detection rule
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RulePatchRequest'
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
  /api/v1/rules/{id}/dry-run:
    post:
      tags: [Rules]
      summary: Dry-run rule against historical events
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Dry-run result
          content:
            application/json:
              schema:
                type: object
                properties:
                  matched_events:
                    type: integer
                  would_create_incident:
                    type: boolean
                required: [matched_events, would_create_incident]
  /api/v1/incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, investigating, resolved, false_positive]
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
                required: [items]
  /api/v1/incidents/{id}:
    get:
      tags: [Incidents]
      summary: Get incident details
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
    patch:
      tags: [Incidents]
      summary: Update incident state
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentPatchRequest'
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
...
  /api/v1/ai/analyze/incident/{id}:
    post:
      tags: [AI]
      summary: Analyze one incident
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                model_profile_id:
                  type: string
                  format: uuid
      responses:
        "202":
          description: Analysis queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobAccepted'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Source:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [file, syslog, journald, docker]
        config:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, type, config, enabled, created_at, updated_at]
    SourceCreateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [file, syslog, journald, docker]
        config:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
          default: true
      required: [name, type, config]
    SourcePatchRequest:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
...
    Rule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        condition:
          type: object
          additionalProperties: true
        threshold:
          type: integer
          minimum: 1
        window_seconds:
          type: integer
          minimum: 1
        severity:
          type: string
        enabled:
          type: boolean
      required: [id, name, condition, threshold, window_seconds, severity, enabled]
    RuleCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          type: object
          additionalProperties: true
        threshold:
          type: integer
          minimum: 1
        window_seconds:
          type: integer
          minimum: 1
        severity:
          type: string
        enabled:
          type: boolean
          default: true
      required: [name, condition, threshold, window_seconds, severity]
    RulePatchRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          type: object
          additionalProperties: true
        threshold:
          type: integer
          minimum: 1
        window_seconds:
          type: integer
          minimum: 1
        severity:
          type: string
        enabled:
          type: boolean
    Incident:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [open, investigating, resolved, false_positive]
        severity:
          type: string
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        event_count:
          type: integer
        rule_id:
          type: string
          format: uuid
          nullable: true
        summary:
          type: string
          nullable: true
        assignee:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required: [id, title, status, severity, first_seen, last_seen, event_count, tags]
    IncidentPatchRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, investigating, resolved, false_positive]
        summary:
          type: string
        assignee:
          type: string
        tags:
          type: array
          items:
            type: string
...
    AsyncJobAccepted:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
      required: [job_id, status]
...